requirements.txt

python-telegram-bot[job-queue]
pymongo
